requirements.txt
fastapi==0.100.0
langchain==0.0.239
numpy==1.23.5
paddleocr==2.6.1.3
paddlepaddle==2.4.2
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.28.1
uvicorn==0.21.1
python-multipart==0.0.6
python-dotenv==1.0.0
openai==0.27.8
